refactor(nn): keep the DCM regularizer internals crate-private
The regularization work widened `ferx-core`'s public API by 17 items, but
only two of them are API: `FitOptions::nn_l2_lambda` and
`nn_smooth_lambda`, which ferx-r sets. The other 15 — `NnRegularizer` and
its four methods, `CurvatureGrid` and its two, and five `MlpMapper`
penalty helpers — are called only from `src/estimation/*`. Their sole
out-of-crate consumer was `tests/nn_regularization.rs`, which is exactly
the "public because a test needs it" case the workspace boundary rule
(CLAUDE.md A3/A6) exists to reject: an item is either public API —
documented, in the baseline, reachable from ferx-r — or it stays
`pub(crate)` and the caller does without.

So they are `pub(crate)` now, and the tests that need them moved into
`src/nn/mod.rs` as `mod regularizer_fit_tests`:

- `regularizer_lambda_zero_is_noop` and
  `regularizer_penalty_gradient_matches_fd` (Tier 1, no fit)
- `l2_shrinks_weights_and_modulator_variation` (Tier 3, `slow-tests`)

`tests/nn_regularization.rs` keeps what genuinely exercises the public
path — `fit_with_regularization_runs` and
`zero_lambda_fit_matches_baseline` drive `nn_l2` / `nn_smooth` through
`fit()` the way ferx-r does and touch no crate-private item, so the file
still proves the feature is usable from outside the crate. It stays in
the endpoint-coverage job's `--test` list; the moved tests are measured
via that job's `--lib`, which the job comment already names as the bulk
of the `nn` surface.

The fixture is duplicated across the two files rather than shared: an
integration test cannot see `#[cfg(test)]` items, so there is nowhere to
put a common copy that both can reach. Both carry a note pointing at the
other.

Net public-API diff against `main` is now:

  +pub ferx_core::types::FitOptions::nn_l2_lambda: f64
  +pub ferx_core::types::FitOptions::nn_smooth_lambda: f64
  +pub ferx_core::FitOptions::nn_l2_lambda: f64
  +pub ferx_core::FitOptions::nn_smooth_lambda: f64

Validation: `tools/preflight.sh` clean; `cargo test --no-default-features
--features ci,markov,nn` 4910 passed / 0 failed (unchanged — the moved
tests are the same tests); under `slow-tests`, `nn::regulariz*` 11/11 and
`nn_regularization` 2/2.
