+++ src/nn/mod.rs
@@ -503,7 +503,7 @@ impl MlpMapper {
     /// `[W_l, b_l]` layout and the parser's `W_…`/`B_…` theta naming). Used by
     /// the L2 (weight-decay) penalty, which shrinks weights but leaves biases
     /// free.
-    pub fn weight_param_indices(&self) -> Vec<usize> {
+    pub(crate) fn weight_param_indices(&self) -> Vec<usize> {
         let mut idx = Vec::new();
         for l in 1..self.layers.len() {
             let start = self.offsets[l - 1];
@@ -515,7 +515,7 @@ impl MlpMapper {
 
     /// L2 (weight-decay) penalty value `lambda · Σ wᵢ²` over the **weight**
     /// blocks only (biases excluded). `lambda == 0.0` returns `0.0` (no-op).
-    pub fn l2_weight_penalty_value(&self, weights: &[f64], lambda: f64) -> f64 {
+    pub(crate) fn l2_weight_penalty_value(&self, weights: &[f64], lambda: f64) -> f64 {
         if lambda == 0.0 {
             return 0.0;
         }
@@ -533,7 +533,7 @@ impl MlpMapper {
     /// L2 penalty value and its gradient w.r.t. the flat weight vector.
     /// `grad[i] = 2·lambda·wᵢ` on weight entries, `0` on bias entries.
     /// `lambda == 0.0` returns `(0.0, zeros)` (strict no-op).
-    pub fn l2_weight_penalty(&self, weights: &[f64], lambda: f64) -> (f64, Vec<f64>) {
+    pub(crate) fn l2_weight_penalty(&self, weights: &[f64], lambda: f64) -> (f64, Vec<f64>) {
         let mut grad = vec![0.0f64; self.n_params];
         if lambda == 0.0 {
             return (0.0, grad);
@@ -555,7 +555,7 @@ impl MlpMapper {
     /// each output along a marginal partial-dependence curve (see
     /// [`CurvatureGrid`]). Forward passes only; `lambda == 0.0` or an empty grid
     /// returns `0.0` (no-op).
-    pub fn curvature_penalty_value(
+    pub(crate) fn curvature_penalty_value(
         &self,
         weights: &[f64],
         grid: &CurvatureGrid,
@@ -583,7 +583,7 @@ impl MlpMapper {
     /// `∂pen/∂w = Σ 2·lambda·C·(∂C/∂w)`. No autodiff needed. `lambda == 0.0` or
     /// an empty grid returns `(0.0, zeros)` (strict no-op). The gradient touches
     /// bias entries too (biases shift the curve), unlike the L2 term.
-    pub fn curvature_penalty(
+    pub(crate) fn curvature_penalty(
         &self,
         weights: &[f64],
         grid: &CurvatureGrid,
@@ -1052,7 +1052,7 @@ pub struct CovariateNn {
 /// z-scored sweep is mapped back to raw values (`z·sd + mean`) here; the raw
 /// vectors are what feed [`MlpMapper::forward`] / [`MlpMapper::jacobian`].
 #[derive(Debug, Clone, Default)]
-pub struct CurvatureGrid {
+pub(crate) struct CurvatureGrid {
     /// Each entry is `[x_minus, x_center, x_plus]`, raw input vectors of length
     /// `n_inputs`.
     triples: Vec<[Vec<f64>; 3]>,
@@ -1066,7 +1066,7 @@ impl CurvatureGrid {
     /// An input with zero spread (constant covariate) or too narrow a range to
     /// fit an interior triple contributes nothing. Returns an empty grid (a
     /// no-op penalty) when no input yields a usable sweep.
-    pub fn build(input_values: &[Vec<f64>], step_z: f64) -> Self {
+    pub(crate) fn build(input_values: &[Vec<f64>], step_z: f64) -> Self {
         let n_in = input_values.len();
         let mut triples = Vec::new();
         if step_z <= 0.0 {
@@ -1131,7 +1131,7 @@ impl CurvatureGrid {
     }
 
     /// Whether the grid contributes any curvature terms.
-    pub fn is_empty(&self) -> bool {
+    pub(crate) fn is_empty(&self) -> bool {
         self.triples.is_empty()
     }
 }
@@ -1175,7 +1175,7 @@ fn collect_input_values(mapper: &NamedMlpMapper, population: &Population) -> Vec
 /// Both penalties are strict no-ops when their λ is `0.0`, so an unregularized
 /// fit adds nothing.
 #[derive(Debug, Clone)]
-pub struct NnRegularizer {
+pub(crate) struct NnRegularizer {
     l2_lambda: f64,
     smooth_lambda: f64,
     specs: Vec<NnRegSpec>,
@@ -1193,7 +1193,11 @@ impl NnRegularizer {
     /// Build the regularizer from the model's `[covariate_nn]` blocks and the
     /// fit options' `nn_l2` / `nn_smooth` strengths. When both λ are `0.0` the
     /// spec list is empty and every method is a no-op.
-    pub fn build(model: &CompiledModel, population: &Population, options: &FitOptions) -> Self {
+    pub(crate) fn build(
+        model: &CompiledModel,
+        population: &Population,
+        options: &FitOptions,
+    ) -> Self {
         let l2_lambda = options.nn_l2_lambda;
         let smooth_lambda = options.nn_smooth_lambda;
         let mut specs = Vec::new();
@@ -1225,14 +1229,14 @@ impl NnRegularizer {
     }
 
     /// Whether any penalty term is active (non-zero λ and at least one NN).
-    pub fn is_active(&self) -> bool {
+    pub(crate) fn is_active(&self) -> bool {
         !self.specs.is_empty() && (self.l2_lambda > 0.0 || self.smooth_lambda > 0.0)
     }
 
     /// Additive penalty on the population objective: `Σ_nn (L2 + smoothness)`.
     /// `theta` is the full (natural-space) theta vector; the NN weight blocks
     /// are sliced out at each spec's offset. Returns `0.0` when inactive.
-    pub fn penalty_value(&self, theta: &[f64]) -> f64 {
+    pub(crate) fn penalty_value(&self, theta: &[f64]) -> f64 {
         let mut p = 0.0;
         for spec in &self.specs {
             let w = &theta[spec.weights_offset..spec.weights_offset + spec.n_weights];
@@ -1253,7 +1257,7 @@ impl NnRegularizer {
     /// optimizer scale, so the raw-weight gradient maps one-to-one into the
     /// packed gradient `grad` (no Jacobian-of-scaling factor). No-op when
     /// inactive.
-    pub fn add_packed_gradient(&self, theta: &[f64], grad: &mut [f64]) {
+    pub(crate) fn add_packed_gradient(&self, theta: &[f64], grad: &mut [f64]) {
         for spec in &self.specs {
             let off = spec.weights_offset;
             let w = &theta[off..off + spec.n_weights];
@@ -2054,5 +2058,270 @@ mod regularization_tests {
             let fd = (vp - vm) / (2.0 * eps);
             assert_relative_eq!(g[j], fd, epsilon = 1e-6, max_relative = 1e-5);
         }
+    }
+}
+
+/// Fit-driven tests for the covariate-NN regularizers.
+///
+/// These live in `src/` rather than `tests/` because they reach for
+/// [`NnRegularizer`] and [`MlpMapper::weight_param_indices`], which are
+/// `pub(crate)`: the estimation layer is their only caller, and the workspace
+/// boundary rule (CLAUDE.md) says an item is either public API — documented, in
+/// `api/ferx-core-public-api.txt`, reachable from ferx-r — or it stays
+/// crate-private and the caller does without. A test is not a reason to widen
+/// the surface. What genuinely exercises the *public* path (driving `nn_l2` /
+/// `nn_smooth` through `fit()` the way ferx-r does) stays in
+/// `tests/nn_regularization.rs`.
+#[cfg(test)]
+mod regularizer_fit_tests {
+    use crate::parser::model_parser::parse_full_model;
+    use crate::types::{CompiledModel, FitOptions, Population};
+    use crate::{fit, read_nonmem_csv};
+
+    use super::{CovariateMapper, NnRegularizer};
+
+    /// Two-cpt oral DCM with a small MLP mapping (WT, CRCL) → the five PK params.
+    ///
+    /// Kept in step with the copy in `tests/nn_regularization.rs`; the two cannot
+    /// share a fixture because an integration test cannot see `#[cfg(test)]` items.
+    ///
+    /// `center` / `scale` are load-bearing, not cosmetic. `WT` runs 45–90 and
+    /// `CRCL` 46–82 on this dataset, so on raw inputs the Glorot-initialised first
+    /// layer produces pre-activations spanning ±170 — `tanh` is clipped at ±1 for
+    /// every subject and the network is constant *by saturation* before the fit
+    /// starts, making the λ = 0 baseline in
+    /// [`l2_shrinks_weights_and_modulator_variation`] flat for the wrong reason.
+    /// z-scoring puts layer 1 in tanh's responsive band, and matches the space the
+    /// `nn_smooth` curvature grid already works in.
+    const MODEL: &str = r#"
+[parameters]
+  omega ETA_CL ~ 0.15
+  omega ETA_V1 ~ 0.15
+  omega ETA_Q  ~ 0.08
+  omega ETA_V2 ~ 0.08
+  omega ETA_KA ~ 0.20
+  sigma PROP_ERR ~ 0.04 (sd)
+
+[covariate_nn TYPICAL_PK]
+  inputs     = [WT, CRCL]
+  outputs    = [CL, V1, Q, V2, KA]
+  layers     = [4]
+  activation = tanh
+  output     = softplus
+  center     = [70, 86]
+  scale      = [12, 23]
+
+[individual_parameters]
+  CL = TYPICAL_PK.CL * exp(ETA_CL)
+  V1 = TYPICAL_PK.V1 * exp(ETA_V1)
+  Q  = TYPICAL_PK.Q  * exp(ETA_Q)
+  V2 = TYPICAL_PK.V2 * exp(ETA_V2)
+  KA = TYPICAL_PK.KA * exp(ETA_KA)
+
+[structural_model]
+  pk two_cpt_oral(cl=CL, v1=V1, q=Q, v2=V2, ka=KA)
+
+[error_model]
+  DV ~ proportional(PROP_ERR)
+
+[fit_options]
+  method = focei
+  maxiter = 50
+  covariance = false
+"#;
+
+    fn load() -> (CompiledModel, FitOptions, Population) {
+        let parsed = parse_full_model(MODEL).expect("model parses with --features nn");
+        let population = read_nonmem_csv(
+            std::path::Path::new("data/two_cpt_oral_cov.csv"),
+            Some(&["WT", "CRCL"]),
+            None,
+        )
+        .expect("dataset loads");
+        (parsed.model, parsed.fit_options, population)
+    }
+
+    /// Variance of the NN's CL modulator (output 0) across subjects, evaluated at
+    /// a given theta. Higher = more covariate-driven spread in the learned curve.
+    fn cl_modulator_variance(model: &CompiledModel, population: &Population, theta: &[f64]) -> f64 {
+        let nn = &model.covariate_nns[0];
+        let n_w = nn.mapper.n_weights();
+        let w = &theta[nn.weights_offset..nn.weights_offset + n_w];
+        let cls: Vec<f64> = population
+            .subjects
+            .iter()
+            .map(|s| nn.mapper.forward_raw(w, &s.covariates).expect("forward")[0])
+            .collect();
+        let n = cls.len() as f64;
+        let mean = cls.iter().sum::<f64>() / n;
+        cls.iter().map(|c| (c - mean) * (c - mean)).sum::<f64>() / n
+    }
+
+    /// L2 norm of the NN's *weight matrices* (biases excluded) at a given theta —
+    /// the quantity `nn_l2` directly shrinks.
+    fn weight_block_sq_norm(model: &CompiledModel, theta: &[f64]) -> f64 {
+        let nn = &model.covariate_nns[0];
+        let w = &theta[nn.weights_offset..nn.weights_offset + nn.mapper.n_weights()];
+        nn.mapper
+            .mlp()
+            .weight_param_indices()
+            .iter()
+            .map(|&i| w[i] * w[i])
+            .sum()
+    }
+
+    /// λ = 0 must leave the regularizer inactive, its penalty exactly 0 and its
+    /// gradient contribution exactly 0 — the no-op guarantee that keeps
+    /// unregularized fits byte-identical.
+    #[test]
+    fn regularizer_lambda_zero_is_noop() {
+        let (model, mut options, population) = load();
+        options.nn_l2_lambda = 0.0;
+        options.nn_smooth_lambda = 0.0;
+        let reg = NnRegularizer::build(&model, &population, &options);
+        assert!(!reg.is_active(), "λ = 0 regularizer must be inactive");
+
+        let theta = model.default_params.theta.clone();
+        assert_eq!(reg.penalty_value(&theta), 0.0, "λ = 0 penalty must be 0");
+
+        let mut grad = vec![0.0; theta.len()];
+        reg.add_packed_gradient(&theta, &mut grad);
+        assert!(
+            grad.iter().all(|&g| g == 0.0),
+            "λ = 0 gradient contribution must be exactly 0"
+        );
+    }
+
+    /// The analytic penalty gradient must match central finite differences of
+    /// `penalty_value` at λ > 0, and must touch only the NN-weight coordinates.
+    #[test]
+    fn regularizer_penalty_gradient_matches_fd() {
+        let (model, mut options, population) = load();
+        options.nn_l2_lambda = 1e-2;
+        options.nn_smooth_lambda = 1e-1;
+        let reg = NnRegularizer::build(&model, &population, &options);
+        assert!(reg.is_active());
+
+        let theta = model.default_params.theta.clone();
+        let mut grad = vec![0.0; theta.len()];
+        reg.add_packed_gradient(&theta, &mut grad);
+
+        // Only the NN-weight coordinates carry a penalty gradient; everything else
+        // (omegas, sigma) must be untouched.
+        let nn = &model.covariate_nns[0];
+        let (w_lo, w_hi) = (nn.weights_offset, nn.weights_offset + nn.mapper.n_weights());
+        for (k, &g) in grad.iter().enumerate() {
+            if k < w_lo || k >= w_hi {
+                assert_eq!(g, 0.0, "non-NN coordinate {k} must have zero penalty grad");
+            }
+        }
+
+        // Central FD of penalty_value w.r.t. each NN weight theta.
+        let eps = 1e-6;
+        let mut p = theta.clone();
+        for k in w_lo..w_hi {
+            let saved = p[k];
+            p[k] = saved + eps;
+            let vp = reg.penalty_value(&p);
+            p[k] = saved - eps;
+            let vm = reg.penalty_value(&p);
+            p[k] = saved;
+            let fd = (vp - vm) / (2.0 * eps);
+            let tol = 1e-6 + 1e-5 * fd.abs();
+            assert!(
+                (grad[k] - fd).abs() <= tol,
+                "penalty grad mismatch at weight {k}: analytic {}, fd {}",
+                grad[k],
+                fd
+            );
+        }
+    }
+
+    /// Growing the L2 strength shrinks the NN weights, flattening the learned
+    /// covariate→modulator curve so the across-subject modulator variance
+    /// collapses toward 0.
+    ///
+    /// The robust, deterministic signal is the **fitted weight-block norm**: L2
+    /// adds `2λw` to the weight gradient, so a heavier λ pulls the optimum's
+    /// weights closer to 0 (monotonically).
+    ///
+    /// The modulator variance is the *effect* being claimed, and it is only a
+    /// meaningful check when the λ = 0 fit actually produces spread to remove. On
+    /// this null-covariate dataset the unregularized fit invents a large spurious
+    /// CL modulator variance (~480 across subjects) — precisely the overfitting
+    /// `nn_l2` exists to suppress — and L2 collapses it to ~0. Asserting both ends
+    /// keeps the oracle non-degenerate in the sense CLAUDE.md requires: an
+    /// assertion that the regularized modulator is flat is worthless if the
+    /// unregularized one was flat too. It was, on raw inputs, because the network
+    /// was saturated rather than because it had learned nothing — see the
+    /// `center` / `scale` note on `MODEL`.
+    #[test]
+    #[cfg_attr(
+        not(feature = "slow-tests"),
+        ignore = "slow: opt in with --features slow-tests"
+    )]
+    fn l2_shrinks_weights_and_modulator_variation() {
+        let (model, options, population) = load();
+
+        let fit_at = |lambda: f64| -> Vec<f64> {
+            let mut o = options.clone();
+            o.nn_l2_lambda = lambda;
+            fit(&model, &population, &model.default_params, &o)
+                .unwrap_or_else(|e| panic!("fit at λ={lambda} failed: {e}"))
+                .theta
+        };
+
+        let t0 = fit_at(0.0);
+        let t_mid = fit_at(5.0);
+        let t_big = fit_at(100.0);
+
+        let (n0, n_mid, n_big) = (
+            weight_block_sq_norm(&model, &t0),
+            weight_block_sq_norm(&model, &t_mid),
+            weight_block_sq_norm(&model, &t_big),
+        );
+        let (v0, v_mid, v_big) = (
+            cl_modulator_variance(&model, &population, &t0),
+            cl_modulator_variance(&model, &population, &t_mid),
+            cl_modulator_variance(&model, &population, &t_big),
+        );
+        eprintln!(
+            "weight ‖W‖²: λ=0 {n0:.5}, λ=5 {n_mid:.5}, λ=100 {n_big:.5}\n\
+             CL modulator var: λ=0 {v0:.6}, λ=5 {v_mid:.6}, λ=100 {v_big:.6}"
+        );
+
+        // Decisive signal: the fitted weight norm shrinks strongly and
+        // monotonically with λ (observed here ~2159 → ~0.004 → ~0.003). This is
+        // the guaranteed mechanism by which L2 flattens the covariate→modulator
+        // map.
+        assert!(
+            n_mid <= n0 + 1e-9 && n_big <= n_mid + 1e-9,
+            "weight norm must be non-increasing in λ (‖W‖²: {n0:.5} → {n_mid:.5} → {n_big:.5})"
+        );
+        assert!(
+            n_big < n0 * 0.5,
+            "heavy L2 (λ=100) must more than halve the fitted weight norm \
+             ({n_big:.5} vs λ=0 {n0:.5})"
+        );
+
+        // The unregularized fit must actually overfit — otherwise the flatness
+        // check below passes against a baseline that was already flat and proves
+        // nothing.
+        assert!(
+            v0 > 1.0,
+            "the λ=0 fit must invent real spurious CL spread for this test to have \
+             a baseline to remove (var {v0:.6}); a near-zero unregularized variance \
+             means the fixture is degenerate, not that L2 worked"
+        );
+
+        // The effect: L2 collapses that spurious spread toward a constant map.
+        // Both regularized fits must be flat; their ordering *relative to each
+        // other* is not asserted, because at ~1e-9 the difference between them is
+        // float noise rather than an effect of λ.
+        assert!(
+            v_mid < 1e-3 && v_big < 1e-3,
+            "L2 must collapse the spurious CL modulator spread \
+             ({v0:.6} → {v_mid:.6} → {v_big:.6})"
+        );
     }
 }
